Add netplotly to p3lib.
This module can be used to manage plotly graphs on a webserver and allow
them to be transferred over a network via ssh.

Plots are saved to a local folder. The local folder maybe served via a web server.
An APi is provided to transfer the local folder contents to a remote machine via ssh.
The contents of the remote folder maybe served via a web server.

The local and remote folders contain an index.html file. This displays a table of
all the available plots. Each plot name in the table maybe clicked which will display
the plotly plot.