Change to make control of the progress bar more flexible.
The progress bar can simply update on a timer every second with durationSeconds set to the expected length
of the task if startMessage is unset and startMessage = None and expectedMessageCount = 0.

Alternatively if durationSeconds is set and startMessage (expectedMessageCount = 0) is set to some text that
we expect to receive in the log before the progress timer (as detailed above) is triggered.

Alternatively if expectedMsgList contains a list of strings we expect to receive then the progress bar is
updated as each message is received. The messages may be the entire line of a log message or parts of a log message line.